+++ packages/link-modules/src/definitions/index.ts
@@ -1,0 +1,3 @@
+export * from "./inventory-level-stock-location"
+export * from "./product-variant-inventory-item"
+export * from "./product-variant-money-amount"