+++ packages/pricing/src/repositories/index.ts
@@ -1,0 +1,3 @@
+export { MikroOrmBaseRepository as BaseRepository } from "@medusajs/utils"
+export { CurrencyRepository } from "./currency"
+export { MoneyAmountRepository } from "./money-amount"