+++ www/api-reference/components/Button/index.tsx
@@ -3,8 +3,8 @@ import clsx from "clsx"
 export type ButtonProps = {
   isSelected?: boolean
   disabled?: boolean
-  variant?: "primary" | "secondary"
-  darkVariant?: "primary" | "secondary"
+  variant?: "primary" | "secondary" | "clear"
+  darkVariant?: "primary" | "secondary" | "clear"
 } & React.HTMLAttributes<HTMLButtonElement>
 
 const Button = ({
@@ -19,8 +19,10 @@ const Button = ({
       className={clsx(
         variant === "primary" && "btn-primary",
         variant === "secondary" && "btn-secondary",
+        variant === "clear" && "btn-clear",
         darkVariant && darkVariant === "primary" && "dark:btn-primary",
         darkVariant && darkVariant === "secondary" && "dark:btn-secondary",
+        darkVariant && darkVariant === "clear" && "dark:btn-clear",
         className
       )}
       {...props}